reflect DrivenPropertyManager until UnityEngine is exposed to ProBuilder
