+++ Runtime/Core/SerializationUtility.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: acc30175e4dd4459ab2fe6953cec7a6f
+guid: c23e2e2798e8047eabd581579b050162
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
